+++ msdk-featdet/msdk-featdet-chromatogrambuilder/src/test/java/io/github/msdk/featdet/chromatogrambuilder/ChromatogramBuilderMethodTest.java
@@ -57,7 +57,7 @@ public class ChromatogramBuilderMethodTest {
     @Test
     public void testOrbitrap() throws MSDKException {
 
-        float noiseLevel = 0f;
+        double noiseLevel = 0;
         double minimumTimeSpan = 6; // 6s
         double minimumHeight = 1E4;
         MZTolerance mzTolerance = new MZTolerance(0.001, 5.0);
@@ -76,7 +76,7 @@ public class ChromatogramBuilderMethodTest {
     public void testBigTimeSpan() throws MSDKException {
 
         // Testing a big timeSpan
-        float noiseLevel = 0f;
+        double noiseLevel = 0;
         double minimumTimeSpan = 1000000;
         double minimumHeight = 1E4;
         MZTolerance mzTolerance = new MZTolerance(0.001, 5.0);
@@ -96,7 +96,7 @@ public class ChromatogramBuilderMethodTest {
     public void testBigMinHeight() throws MSDKException {
 
         // Testing a big minimum height
-        float noiseLevel = 0f;
+        double noiseLevel = 0f;
         double minimumTimeSpan = 6; // 6s
         double minimumHeight = 10000000;
         MZTolerance mzTolerance = new MZTolerance(0.001, 5.0);
